requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
APScheduler==3.10.4
asttokens==2.4.1
astunparse==1.6.3
attrs==25.3.0
babel==2.16.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
check-manifest==0.50
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
coverage==7.6.6
cycler==0.12.1
decorator==5.1.1
deep-sort-realtime==1.3.2
easydict==1.13
executing==2.1.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
gast==0.6.0
gdown==5.2.0
ghp-import==2.1.0
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
griffe==1.5.1
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
iniconfig==2.0.0
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-autorefs==1.2.0
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.44
mkdocs-material-extensions==1.3.1
mkdocstrings==0.27.0
mkdocstrings-python==1.12.2
ml-dtypes==0.4.1
MouseInfo==0.1.3
mpmath==1.3.0
mysql-connector-python==9.1.0
namex==0.0.8
networkx==3.4.2
numpy==2.0.2
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
outcome==1.3.0.post0
packaging==24.1
paginate==0.5.7
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
proto-plus==1.26.1
protobuf==5.28.3
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.10.0
pymdown-extensions==10.12
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyparsing==3.2.0
pyperclip==1.9.0
pyproject_hooks==1.2.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.3.3
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.0
python-socketio==5.13.0
python-telegram-bot==21.7
pytweening==1.2.0
pytz==2024.2
pywhatkit==5.4
PyYAML==6.0.2
pyyaml_env_tag==0.1
regex==2024.11.6
requests==2.32.3
rich==13.9.3
rsa==4.9.1
safetensors==0.6.2
schedule==1.2.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.31.0
setuptools==75.2.0
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
thop==0.1.1.post2209072238
tokenizers==0.21.4
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.55.0
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
# Editable install with no version control (ultralytics==8.3.31)
-e c:\users\danny\appdata\local\programs\python\python312\lib\site-packages
ultralytics-thop==2.0.11
uritemplate==4.2.0
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.5
wheel==0.44.0
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
